requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
packaging==23.2
pillow==10.2.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1

watchfiles==0.21.0
websockets==12.0
